labeling/requirements.txt
asyncio
aiohttp
python-dateutil
atproto
textblob
pyyaml